refactor: break down storage_details helpers into utils
