+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/storage_details/storage_treemap/index.tsx
@@ -22,12 +22,6 @@ import { EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import React, { useCallback, useMemo } from 'react';
 
 import { i18n } from '@kbn/i18n';
-import {
-  FlattenedBucket,
-  getLayersMultiDimensional,
-  getLegendItems,
-  getPathToFlattenedBucketMap,
-} from '../helpers';
 import { ChartLegendItem } from './chart_legend_item';
 import { NoData } from './no_data';
 import {
@@ -36,6 +30,10 @@ import {
 } from '../../indices_details/pattern/index_check_flyout/index_properties/index_check_fields/tabs/styles';
 import { PatternRollup, SelectedIndex } from '../../../types';
 import { useDataQualityContext } from '../../../data_quality_context';
+import { FlattenedBucket } from '../types';
+import { getPathToFlattenedBucketMap } from './utils/get_path_to_flattened_bucket_map';
+import { getLayersMultiDimensional } from './utils/get_layers_multi_dimensional';
+import { getLegendItems } from './utils/get_legend_items';
 
 export const DEFAULT_MIN_CHART_HEIGHT = 240; // px
 export const LEGEND_WIDTH = 220; // px
