+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/data_quality_details/storage_details/storage_treemap/utils/get_path_to_flattened_bucket_map.ts
@@ -1,0 +1,19 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { FlattenedBucket } from '../../types';
+
+export const getPathToFlattenedBucketMap = (
+  flattenedBuckets: FlattenedBucket[]
+): Record<string, FlattenedBucket | undefined> =>
+  flattenedBuckets.reduce<Record<string, FlattenedBucket | undefined>>(
+    (acc, { pattern, indexName, ...remaining }) => ({
+      ...acc,
+      [`${pattern}${indexName}`]: { pattern, indexName, ...remaining },
+    }),
+    {}
+  );
