+++ x-pack/packages/security-solution/ecs_data_quality_dashboard/impl/data_quality_panel/hooks/use_results_rollup/utils/get_pattern_rollups_with_latest_check_result.ts
@@ -5,11 +5,11 @@
  * 2.0.
  */
 
-import { getIndexDocsCountFromRollup } from '../../../data_quality_summary/summary_actions/check_all/helpers';
 import { getIlmPhase } from '../../../data_quality_details/indices_details/pattern/helpers';
 import { getAllIncompatibleMarkdownComments } from '../../../data_quality_details/indices_details/pattern/index_check_flyout/index_properties/index_check_fields/tabs/incompatible_tab/helpers';
 import { getSizeInBytes } from '../../../utils/stats';
 import type { IlmPhase, PartitionedFieldMetadata, PatternRollup } from '../../../types';
+import { getIndexDocsCountFromRollup } from './stats';
 
 export const getPatternRollupsWithLatestCheckResult = ({
   error,
